Fixed Emulator Issue
When I downloaded the new application from our Github, it wasn't displaying the pictures or anything. I took the time to correct these issues to make sure that our MainActivity background.jpg was working properly, and well as any other image that may have not shown up at the time.